move filter objects module to classes folder
