+++ pyfar/classes/__init__.py
@@ -1,10 +1,13 @@
 from .signal import Signal, TimeData, FrequencyData
 from .coordinates import Coordinates
 from .orientations import Orientations
+from .filter import FilterFIR, FilterIIR, FilterSOS
 
 from . import signal
 from . import coordinates
 from . import orientations
+from . import filter
+
 
 __all__ = [
     'signal',
@@ -15,4 +18,8 @@ __all__ = [
     'FrequencyData',
     'Coordinates',
     'Orientations',
+    'FilterFIR',
+    'FilterIIR',
+    'FilterSOS',
+    'filter'
 ]
