+++ pyfar/classes/filter.py
