+++ pyfar/io.py
@@ -21,7 +21,7 @@ import io
 from pyfar import Signal
 from pyfar import Coordinates
 import pyfar._codec as codec
-import pyfar.dsp.classes as fo
+import pyfar.classes.filter as fo
 
 
 def read_wav(filename):
